Resource key support for regular and singleton resources.
commit 07b476913520609becaec30878206c51c179a593
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Thu Aug 13 08:35:13 2015 +0200

    docs and comments

commit 2d2bb136640e84707dd315d6e67b7b2c3d972717
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Thu Aug 13 08:33:14 2015 +0200

    cleanup

commit bf4e642991ce31da49b5ca3749d887258a5a1a93
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 13:20:05 2015 +0200

    fix integration test case

commit 9e6dcb2850c10f9f8f84b2e4d575670e52108eb9
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 11:58:05 2015 +0200

    add fields for singleton subresources

commit 68ea1b12f56de955c9b386cd295fa5768c1e2714
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 11:30:44 2015 +0200

    generate sub resources into their own (sub) package

commit 0b3cc7752aa719c5bb6acb53f7078a48d9737a7b
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 10:30:05 2015 +0200

    cleanup: remove unused classes

commit e70b5b7e1272c683c42d09c85501358253a983b2
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 10:25:39 2015 +0200

    remove extra operation to retrieve the singleton meta data

commit 83551f0be423cbf264619b2ab1a1a8c1d2fa9c8b
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 09:09:22 2015 +0200

    checkpoint: iplicit resource names for singleton resources

commit 21769983b6065587b1b37a498307230f8ec5b2ca
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Wed Aug 12 08:55:00 2015 +0200

    fix child description condition

commit a08004e8fe29d07ecc349d5f5583a8f5a9ec72cb
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 16:50:21 2015 +0200

    Illegal state when resource description is undefined

commit 40aa7f9a9d23b20001f3b665f1aa877aefebf372
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 16:26:04 2015 +0200

    Checkpoint: ctor invocation for keyset enums

commit 55d3a94d384b94a44fd41406d6d374e63c462c23
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 16:20:48 2015 +0200

    fix test coverage

commit 780a9f41b8f345492c00e1b9605d0079d00452a0
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 16:20:22 2015 +0200

    instantiate inner classes correctly

commit 4bfed24840ccf805b994f6814143c03cceef97ca
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 16:09:44 2015 +0200

    Error handling improvements

commit 4dc1347916bfb96d274fb811dd2381064bcb04e1
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 15:24:57 2015 +0200

    (WIP): singleton resources are not properly parsed: the attributes are missing

commit d85ab9f2e27742affa9795eebcf47425eec259b3
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 15:06:34 2015 +0200

    support keys as enums

commit 6743067e3030df3512798c34ece86ab282459577
Author: Heiko Braun <ike.braun@googlemail.com>
Date:   Tue Aug 11 12:43:24 2015 +0200

    added sub resources singleton meta data

+++ generator/src/main/java/org/wildfly/apigen/generator/Generator.java
@@ -5,18 +5,25 @@ import org.jboss.dmr.ModelNode;
 import org.jboss.forge.roaster.model.source.JavaClassSource;
 import org.jboss.logmanager.Level;
 import org.wildfly.apigen.invocation.ClientFactory;
+import org.wildfly.apigen.model.AddressTemplate;
 import org.wildfly.apigen.model.DefaultStatementContext;
 import org.wildfly.apigen.model.ResourceDescription;
 import org.wildfly.apigen.operations.ReadDescription;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.SimpleFileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.logging.Logger;
+
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.*;
 
 /**
  * @author Heiko Braun
@@ -24,14 +31,16 @@ import java.util.logging.Logger;
  */
 public class Generator {
 
+    private static final Logger log = Logger.getLogger(Generator.class.getName());
     private static ModelControllerClient client;
 
-    private static final Logger log = Logger.getLogger(Generator.class.getName());
+    private DefaultStatementContext statementContext;
 
     private final String targetDir;
     private final Config config;
 
     public Generator(String targetDir, Config config) {
+        this.statementContext = new DefaultStatementContext();
         this.targetDir = targetDir;
         this.config = config;
 
@@ -47,29 +56,68 @@ public class Generator {
         log.info("Config: " + args[0]);
         log.info("Output: " + args[1]);
 
-        Config config = Config.fromJson(args[0]);
-        Generator generator = new Generator(args[1], config);
-        generator.processGeneratorTargets();
-        generator.shutdown();
+        try {
+            Config config = Config.fromJson(args[0]);
+            Generator generator = new Generator(args[1], config);
+            generator.processGeneratorTargets();
+            generator.shutdown();
+        } catch (Throwable e) {
+            System.exit(-1);
+            log.log(Level.ERROR, "Unexpected error", e);
+        }
     }
 
-    private void shutdown() {
+    public void deleteDir(String dir) throws Exception
+    {
+        Path directory = Paths.get(dir);
+        Files.walkFileTree(directory, new SimpleFileVisitor<Path>() {
+            @Override
+            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                Files.delete(file);
+                return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+                Files.delete(dir);
+                return FileVisitResult.CONTINUE;
+            }
+
+        });
+    }
+
+
+    public void shutdown() {
         try {
             client.close();
         } catch (IOException e) {
-            log.severe(e.getMessage());
+            log.log(Level.ERROR, e.getMessage());
         }
     }
 
     public void processGeneratorTargets() {
+
+
+        if(Files.exists(Paths.get(targetDir))) {
+            System.out.println("Delete output dir: "+ targetDir);
+            try {
+                deleteDir(targetDir);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+
         config.getGeneratorTargets().forEach(
                 target -> {
                     try {
                         GeneratorScope scope = new GeneratorScope();
-                        ResourceMetaData resourceMetaData = fetchMetaData(target);
+
+                        // load resource entry point recursively
+                        ResourceMetaData resourceMetaData = loadResourceMetaData(target);
                         resourceMetaData.set(ResourceMetaData.PKG, target.getTargetPackage());
 
-                        Iterator<ResourceMetaData> iterator = new MetaDataIterator(resourceMetaData).createInstance();
+                        // generate classes
+                        Iterator<ResourceMetaData> iterator = new MetaDataTraversal(resourceMetaData).createInstance();
                         iterator.forEachRemaining(metaData -> {
                             generate(scope, metaData, targetDir);
                         });
@@ -82,20 +130,62 @@ public class Generator {
         );
     }
 
-    private static ResourceMetaData fetchMetaData(GeneratorTarget generatorTarget) throws Exception {
-        ReadDescription op = new ReadDescription(generatorTarget.getSourceAddress());
-        ModelNode response = client.execute(op.resolve(new DefaultStatementContext()));
-        return new ResourceMetaData(generatorTarget.getSourceAddress(), ResourceDescription.from(response));
+    private ResourceMetaData loadResourceMetaData(GeneratorTarget generatorTarget) throws Exception {
+
+        AddressTemplate address = generatorTarget.getSourceAddress();
+
+        ModelNode composite = new ModelNode();
+        composite.get(OP).set(COMPOSITE);
+        composite.get(ADDRESS).setEmptyList();
+
+        List<ModelNode> steps = new ArrayList<>();
+
+        // read parent child type
+        Integer tokens = address.tokenLength();
+        AddressTemplate parentAddress = address.subTemplate(0, tokens - 1);
+        ModelNode childTypes = new ModelNode();
+        childTypes.get(OP).set(READ_CHILDREN_TYPES_OPERATION);
+        childTypes.get(ADDRESS).set(parentAddress.resolve(new DefaultStatementContext()));
+        childTypes.get("include-singletons").set(true);
+        steps.add(childTypes);
+
+        // read resource description
+        ReadDescription rrd = new ReadDescription(address);
+        steps.add(rrd.resolve(this.statementContext));
+
+        composite.get(STEPS).set(steps);
+        ModelNode response = client.execute(composite);
+
+        // parent type
+        boolean isSingleton = false;
+        List<ModelNode> types = response.get(RESULT).get("step-1").get(RESULT).asList();
+        for (ModelNode type : types) {
+            if(type.asString().equals(address.getResourceType()+"="+address.getResourceName())) {
+                isSingleton = true;
+                break;
+            }
+        }
+
+        // resource meta data
+        ResourceDescription description = ResourceDescription.from(response.get(RESULT).get("step-2"));
+        if(isSingleton)
+        {
+            description.setSingletonName(address.getResourceName());
+        }
+
+        return new ResourceMetaData(generatorTarget.getSourceAddress(), description);
     }
 
-    private void generate(
-            GeneratorScope scope, ResourceMetaData resourceMetaData,
-            String targetDir){
+    private void generate(GeneratorScope scope, ResourceMetaData resourceMetaData, String targetDir){
 
-        JavaClassSource javaClass = SourceFactory.createResourceAsClass(resourceMetaData);
+        JavaClassSource javaClass = SourceFactory.createResourceAsClass(scope, resourceMetaData);
 
-        if (!resourceMetaData.getDescription().getChildrenNames().isEmpty()) {
+        if (!resourceMetaData.getDescription().getChildrenTypes().isEmpty()) {
             SourceFactory.createChildAccessors(scope, resourceMetaData, javaClass);
+        }
+
+        if (!resourceMetaData.getDescription().getSingletonChildrenTypes().isEmpty()) {
+            SourceFactory.createSingletonChildAccessors(scope, resourceMetaData, javaClass);
         }
 
         writeClass(targetDir, javaClass);
@@ -109,8 +199,12 @@ public class Generator {
             Files.createDirectories(Paths.get(targetDir));
 
             Path fileName = Paths.get(targetDir + File.separator + javaClass.getName() + ".java");
-            log.info(javaClass.getCanonicalName());
+            if(Files.exists(fileName)) {
+                throw new RuntimeException("File already exists, will be replaced: " + fileName);
+            }
+
             Files.write(fileName, javaClass.toString().getBytes());
+
         } catch (IOException e) {
             log.log(Level.ERROR, "Failed to persist class", e);
         }
